requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
attrs==21.4.0
Django==3.2.14
django-extensions==3.2.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
promise==2.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytz==2022.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
Werkzeug==2.1.2
